\name{selectiongain-package}
\alias{selectiongain-package}
\alias{selectiongain}
\docType{package}
\title{
gain of selection }
\description{

This package calculate the gain from a selection.  It uses part of Tallis' algorithm (1961). 

The original algorithm is to calculate the moment generating function of the truncated Mvtnorm distribution.

There will be a Mvtmoment package for the whole Tallis' algorithm.

}
\details{
\tabular{ll}{
Package: \tab selectiongain\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-11-24\cr
License: \tab GPL \cr
}
the most important function is "multistageselection"
}
\author{
Xuefei Mi, Friedrich Utz
Maintainer:Xuefei Mi
<mi@pz.uni-hohenheim.de>
}
\references{

}
\keyword{ package }
\seealso{
}
\examples{
}
